Initial commit
This is a long script, but it's just doing the same thing for each carbon cohort: First it takes those plots with carbon data and uses RF to predict the carbon values. Then, for those plots without carbon attributes it uses the proper RF model to impute the value corresponding to the most similar plot.